+++ packages/logging/service.go
@@ -42,33 +42,6 @@ type doguDescriptionGetter interface {
 	GetCurrent(ctx context.Context, simpleDoguName string) (*core.Dogu, error)
 }
 
-type LogLevel int
-
-const (
-	LevelUnknown LogLevel = iota
-	LevelError
-	LevelWarn
-	LevelInfo
-	LevelDebug
-)
-
-func (l LogLevel) String() string {
-	switch l {
-	case LevelUnknown:
-		return "UNKNOWN"
-	case LevelDebug:
-		return "DEBUG"
-	case LevelInfo:
-		return "INFO"
-	case LevelWarn:
-		return "WARN"
-	case LevelError:
-		return "ERROR"
-	default:
-		return "WARN"
-	}
-}
-
 // NewLoggingService creates a new logging service.
 func NewLoggingService(provider logProvider, cp configProvider, restarter doguRestarter, descriptionGetter doguDescriptionGetter) *loggingService {
 	return &loggingService{
@@ -140,30 +113,37 @@ func (s *loggingService) GetForDogu(request *pb.DoguLogMessageRequest, server pb
 	return writeLogLinesToStream(s.logProvider, doguName, linesCount, server)
 }
 
-func (s *loggingService) SetLogLevel(ctx context.Context, req *pb.LogLevelRequest) (*emptypb.Empty, error) {
+// ApplyLogLevelWithRestart sets the log level for a specific dogu and restarts the dogu if the log level was changed.
+func (s *loggingService) ApplyLogLevelWithRestart(ctx context.Context, req *pb.LogLevelRequest) (res *emptypb.Empty, err error) {
 	doguName := req.DoguName
 
+	createInternalErrWithCtx := wrapCreateInternalErrWithContext(fmt.Sprintf("error occurred in ApplyLogLevelWithRestart for dogu \"%s\"", doguName))
+
 	if strings.TrimSpace(doguName) == "" {
-		return nil, createInternalErr(errMissingDoguName, codes.InvalidArgument)
+		return nil, createInternalErrWithCtx(errMissingDoguName, codes.InvalidArgument)
 	}
 
-	lLevel, err := mapLogLevelFromProto(req.GetLogLevel())
+	lLevel, err := CreateLogLevelFromProto(req.GetLogLevel())
 	if err != nil {
-		return nil, createInternalErr(fmt.Errorf("unable to map log level from proto message: %w", err), codes.InvalidArgument)
+		return nil, createInternalErrWithCtx(fmt.Errorf("unable to map log level from proto message: %w", err), codes.InvalidArgument)
 	}
 
 	restart, err := s.setLogLevel(ctx, doguName, lLevel)
 	if err != nil {
-		return nil, createInternalErr(fmt.Errorf("unable to set log level: %w", err), codes.Internal)
+		return nil, createInternalErrWithCtx(fmt.Errorf("unable to set log level: %w", err), codes.Internal)
 	}
+
+	logrus.Debugf("restart needed for log level change: %v", restart)
 
 	if !restart {
 		return &emptypb.Empty{}, nil
 	}
 
-	if lErr := s.doguRestarter.RestartDogu(ctx, doguName); lErr != nil {
-		return nil, createInternalErr(fmt.Errorf("unable to restart dogu %s after setting new log level: %w", doguName, lErr), codes.Internal)
+	if lErr := s.doguRestarter.RestartDogu(context.WithoutCancel(ctx), doguName); lErr != nil {
+		return nil, createInternalErrWithCtx(fmt.Errorf("unable to restart dogu %s after setting new log level: %w", doguName, lErr), codes.Internal)
 	}
+
+	logrus.Debugf("Restarted dogu %s", doguName)
 
 	return &emptypb.Empty{}, nil
 }
@@ -184,9 +164,15 @@ func (s *loggingService) setLogLevel(ctx context.Context, doguName string, l Log
 		return false, fmt.Errorf("could not change log level from %s to %s: %w", currentLogLevel, l.String(), err)
 	}
 
+	logrus.Debugf("written new log level %s for dogu %s", l.String(), doguName)
+
 	return true, nil
 }
 
+// GetLogLevel provides the log level currently set for a specific dogu.
+// For this the dogu config is checked as well as the dogu description as fall back.
+// When there is no value set for the log level LevelUnknown is returned without an error.
+// An error is only returned in case dogu config or dogu description cannot be read.
 func (s *loggingService) GetLogLevel(ctx context.Context, doguName string) (LogLevel, error) {
 	dConfig := s.configProvider.DoguConfig(doguName)
 
@@ -200,6 +186,8 @@ func (s *loggingService) getLogLevel(ctx context.Context, doguName string, doguC
 	}
 
 	if currentLogLevelStr == "" {
+		logrus.Debugf("config log level is empty, try to get default log level from dogu descrption")
+
 		currentLogLevelStr, err = s.getDefaultLogLevel(ctx, doguName)
 		if err != nil {
 			return LevelUnknown, fmt.Errorf("could not get default log level from dogu description: %w", err)
@@ -211,7 +199,9 @@ func (s *loggingService) getLogLevel(ctx context.Context, doguName string, doguC
 		return LevelUnknown, nil
 	}
 
-	currentLogLevel, err := mapLogLevelFromString(currentLogLevelStr)
+	logrus.Debugf("current log level from dogu %s is %s", doguName, currentLogLevelStr)
+
+	currentLogLevel, err := CreateLogLevelFromString(currentLogLevelStr)
 	if err != nil {
 		logrus.Warnf("invalid log level set for dogu %s: %s", doguName, currentLogLevelStr)
 
@@ -246,38 +236,6 @@ func (s *loggingService) getDefaultLogLevel(ctx context.Context, doguName string
 	}
 
 	return defaultLevelStr, nil
-}
-
-func mapLogLevelFromProto(pLevel pb.LogLevel) (LogLevel, error) {
-	switch pLevel {
-	case pb.LogLevel_DEBUG:
-		return LevelDebug, nil
-	case pb.LogLevel_INFO:
-		return LevelInfo, nil
-	case pb.LogLevel_WARN:
-		return LevelWarn, nil
-	case pb.LogLevel_ERROR:
-		return LevelError, nil
-	default:
-		return LevelUnknown, fmt.Errorf("unknown log level: %v", pLevel)
-	}
-}
-
-func mapLogLevelFromString(sLevel string) (LogLevel, error) {
-	sLevelUpper := strings.ToUpper(sLevel)
-
-	switch sLevelUpper {
-	case LevelError.String():
-		return LevelError, nil
-	case LevelWarn.String():
-		return LevelWarn, nil
-	case LevelInfo.String():
-		return LevelInfo, nil
-	case LevelDebug.String():
-		return LevelDebug, nil
-	default:
-		return LevelUnknown, errors.New("unknown log level")
-	}
 }
 
 func (s *loggingService) writeLogLevel(_ context.Context, dConfig registry.ConfigurationContext, l LogLevel) error {
@@ -351,4 +309,12 @@ func compressMessages(doguName string, logLines []logLine) ([]byte, error) {
 func createInternalErr(err error, code codes.Code) error {
 	logrus.Error(err)
 	return status.Error(code, err.Error())
+}
+
+func wrapCreateInternalErrWithContext(errCtx string) func(error, codes.Code) error {
+	return func(err error, code codes.Code) error {
+		contextError := fmt.Errorf("%s : %w", errCtx, err)
+
+		return createInternalErr(contextError, code)
+	}
 }