+++ CHANGELOG.md
@@ -9,6 +9,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 - [#27] API GetDoguList reports current LogLevel
 - [#25] API to set log level for a dogu
+  - when the provided log level differs from the log level set in the config, a dogu restart is triggered to apply the change
 
 ## [v0.7.0] - 2024-05-02
 ### Added